requirements.txt
# Copyright 2020-2024 Datum Technology Corporation
# All rights reserved.
#######################################################################################################################



docutils~=0.19
setuptools~=65.5.0
requests~=2.31.0
pydantic~=2.9.2
semantic-version~=2.10.0
PyYAML~=6.0
Jinja2~=3.1.2
toml~=0.10.2
pydantic-extra-types==2.9.0